move abci query router to a module and explain it
